Uploaded First Utilities
These are python utilities for creating an index of the top 2000 crypto currencies, making an assignment file out of the index (even though that didn't exactly get used for its full purpose, but it can be. Right now it is used as a state file for starting the script where it last left off downloading historical OHLCV data for 2000 cryptos for the past 5 years.), and downloading the historical OHLCV data for 2000 top cryptos for the past 5 years. I also uploaded a requirements.txt file from pip freeze to install the necessary libraries for these scripts and more.